[#23] feat: 아이 회원가입시 파킹통장 생성 및 기본 주식 셋팅 기능 추가

+++ backend/src/main/java/com/a502/backend/application/entity/Parking.java
@@ -36,7 +36,7 @@ public class Parking extends BaseEntity {
 	private User user;
 
 	@Builder
-	public Parking(int balance, int interest, User user) {
+	public Parking(int balance, double interest, User user) {
 		this.balance = balance;
 		this.interest = interest;
 		this.user = user;
